backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
supabase>=2.3.0
reportlab>=4.4.5
pillow>=12.0.0
pydantic[email]>=2.10.0
python-multipart>=0.0.6
pymupdf>=1.24.0
email-validator>=2.0.0
